requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.1.1
pandas==1.4.2
openpyxl==3.0.9
python-dotenv==0.19.2 